implement checkpoint egl/gl call order list.
A checkpoint is an indication of the order of gl/egl calls
from different threads (in the future from different processes).
The server keeps a queue of checkpoints.  The client generates
a checkpoint log message when it encounters egl/gl calls that
must be kept in the order.  These include create display/surface/
context, destroy display/surface/context, make current, any
flush/finish, (un)binding surface to image/texture.

When a client encounters such calls, it first send a log message to
the server.  The server then puts it the queue.  The client then
sends the the call through command buffer.  When the server
receives from the command buffer, it checks whether this command
matches the head of the queue.  If it does not match, the server waits,
otherwise, it removes the queue head and proceeds.

+++ src/server/server_state.h
@@ -39,7 +39,7 @@ typedef struct _server_display_list {
 
 typedef struct _server_log {
     thread_t            server;
-    long                timestamp;
+    double              timestamp;
 } server_log_t;
 
 /*********************************************************
@@ -110,19 +110,19 @@ _server_display_find (EGLDisplay egl_display);
  * any other egl/gl calls immediately after these calls
  ********************************************************/
 private void
-_call_order_list_append (thread_t server, long timestamp);
+_call_order_list_append (thread_t server, double timestamp);
 
 private void
 _call_order_list_remove ();
 
 private bool
-_call_order_list_head_is_server (thread_t server, long timestamp);
+_call_order_list_head_is_server (thread_t server, double timestamp);
 
 private void
-_server_append_call_log (thread_t server, long timestamp);
+_server_append_call_log (thread_t server, double timestamp);
 
 private bool
-_server_allow_call (thread_t server, long timestamp);
+_server_allow_call (thread_t server, double timestamp);
 
 private void
 _server_remove_call_log ();
